Bug correction with signalArrival
	- sizeof() returns a size_t, which is unsigned, thus cannot be compared
to an int. in signalArrivalDeparture sizeof is now cast into an int

+++ src/signalArrival.c
@@ -47,7 +47,7 @@ static void signalArrivalDepartures(char typ, address ad, addressSet circuit){
   // We must send a msg even if the wagon is full. Thus, if there is not enough
   // space in the wagon, we realloc it.
   if ((wagonMaxLen - wagonToSend->p_wagon->header.len)
-      < sizeof(payloadArrivalDeparture)) {
+      < (int) sizeof(payloadArrivalDeparture)) {
     int newWomimLen = sizeof(prefix) + wagonToSend->p_wagon->header.len
         + sizeof(messageHeader) + sizeof(payloadArrivalDeparture);
     wagonToSend->p_womim = realloc(wagonToSend->p_womim, newWomimLen);
